fix: overrite or not found chave.key

+++ secure_pass4/crypt_pass4.py
@@ -3,18 +3,27 @@ from cryptography.fernet import Fernet
 
 def gerar_chave():
     """Gera e salva uma chave para criptografia/descriptografia."""
-    chave = Fernet.generate_key()
-    with open("chave.key", "wb") as chave_file:
-        chave_file.write(chave)
+    # attention to do not overwrite chave.key if it exist.
+    # rename or delete it before creating a new key
+    if not os.path.exists("chave.key"):
+        chave = Fernet.generate_key()
+        with open("chave.key", "wb") as chave_file:
+            chave_file.write(chave)
 
 def carregar_chave():
     """Carrega a chave usada para criptografar/descriptografar as senhas."""
+    # return if key not loaded avoing execution error
+    if not os.path.exists("chave.key"):
+        print("Your chave.key File NOT FOUND. PLEASE RECOVERY IT.")
+        return
     with open("chave.key", "rb") as chave_file:
         return chave_file.read()
 
 def criptografar_senha(senha):
     """Criptografa uma senha."""
-    senha_criptografada = ""
+    chave = carregar_chave()
+    fernet = Fernet(chave)
+    senha_criptografada = fernet.encrypt(senha.encode())
     return senha_criptografada
 
 def descriptografar_senha(senha_criptografada):
@@ -23,6 +32,9 @@ def descriptografar_senha(senha_criptografada):
     return senha
 
 
-gerar_chave()
-mykey = carregar_chave()
-print(mykey)
+#gerar_chave()
+#mykey = carregar_chave()
+#if(mykey):
+#    mykey = mykey.decode('utf-8')  # convert from bytes to string before print b``
+#    print(f"{mykey}")
+
